pyproject.toml
[tool.poetry]
name = "forecasting-are"
version = "0.1.0"
description = ""
authors = ["Eli Lifland <eli.d.lifland@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"
numpy = "^2.2.3"
pandas = "^2.2.3"
scipy = "^1.15.2"
matplotlib = "^3.10.1"
tqdm = "^4.67.1"
python-dateutil = "^2.9.0.post0"
streamlit = "^1.46.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
